# Use vectorization for categorical distance