--- a/optuna/samplers/_tpe/parzen_estimator.py
+++ b/optuna/samplers/_tpe/parzen_estimator.py
@@ -187,7 +187,7 @@ class _ParzenEstimator:
     ) -> _BatchedDistributions:
         choices = search_space.choices
         n_choices = len(choices)
-        n_samples = len(observations) + (parameters.consider_prior or len(observations) == 0)
+        n_samples = observations.size + (parameters.consider_prior or observations.size == 0)
 
         assert parameters.prior_weight is not None
         weights = np.full(
@@ -195,16 +195,20 @@ class _ParzenEstimator:
             fill_value=parameters.prior_weight / n_samples,
         )
 
+        observed_indices = observations.astype(int)
         if param_name in parameters.categorical_distance_func:
+            # TODO(nabenabe0928): Think about how to handle combinatorial explosion.
+            # The time complexity is O(n_choices * used_indices.size), so n_choices cannot be huge.
+            used_indices, rev_indices = np.unique(observed_indices, return_inverse=True)
             dist_func = parameters.categorical_distance_func[param_name]
-            dists = np.array([[dist_func(c1, c2) for c1 in choices] for c2 in choices])
+            dists = np.array([[dist_func(choices[i], c) for c in choices] for i in used_indices])
             max_dists = np.max(dists, axis=1)
             coef = np.log(n_samples / parameters.prior_weight) * np.log(n_choices) / np.log(6)
             categorical_weights = np.exp(-dists / max_dists[:, np.newaxis] * coef)
             categorical_weights /= np.sum(categorical_weights, axis=1, keepdims=True)
-            weights = categorical_weights[observations.astype(int)]
+            weights = categorical_weights[rev_indices]
         else:
-            weights[np.arange(len(observations)), observations.astype(int)] += 1
+            weights[np.arange(observed_indices.size), observed_indices] += 1
 
         weights /= weights.sum(axis=1, keepdims=True)
         return _BatchedCategoricalDistributions(weights)