--- a/optuna/samplers/_tpe/parzen_estimator.py
+++ b/optuna/samplers/_tpe/parzen_estimator.py
@@ -187,16 +187,16 @@ class _ParzenEstimator:
     ) -> _BatchedDistributions:
         choices = search_space.choices
         n_choices = len(choices)
-        n_samples = observations.size + (parameters.consider_prior or observations.size == 0)
+        n_kernels = len(observations) + (parameters.consider_prior or len(observations) == 0)
 
         assert parameters.prior_weight is not None
         weights = np.full(
-            shape=(n_samples, n_choices),
-            fill_value=parameters.prior_weight / n_samples,
+            shape=(n_kernels, n_choices),
+            fill_value=parameters.prior_weight / n_kernels,
         )
 
         observed_indices = observations.astype(int)
-        if observed_indices.size == 0:
+        if len(observed_indices) == 0:
             # Do nothing.
             pass
         elif param_name in parameters.categorical_distance_func:
@@ -206,11 +206,11 @@ class _ParzenEstimator:
             dist_func = parameters.categorical_distance_func[param_name]
             dists = np.array([[dist_func(choices[i], c) for c in choices] for i in used_indices])
             max_dists = np.max(dists, axis=1)
-            coef = np.log(n_samples / parameters.prior_weight) * np.log(n_choices) / np.log(6)
+            coef = np.log(n_kernels / parameters.prior_weight) * np.log(n_choices) / np.log(6)
             categorical_weights = np.exp(-((dists / max_dists[:, np.newaxis]) ** 2) * coef)
-            weights[: observed_indices.size] = categorical_weights[rev_indices]
+            weights[: len(observed_indices)] = categorical_weights[rev_indices]
         else:
-            weights[np.arange(observed_indices.size), observed_indices] += 1
+            weights[np.arange(len(observed_indices)), observed_indices] += 1
 
         weights /= weights.sum(axis=1, keepdims=True)
         return _BatchedCategoricalDistributions(weights)