@contramundum53 2023-12-20
Nit: Is there any reason you prefer `observed_indices.size` over `len(observations)`?
This is just a matter of taste, but Optuna developers seems to have previously preferred the latter.