@nabenabe0928 2023-12-22
> Nit: Is there any reason you prefer `observed_indices.size` over `len(observations)`? This is just a matter of taste, but Optuna developers seems to have previously preferred the latter.

The reason why I take `np.size` rather than `len` is that `np.size` imposes a stricter constraint on the input.
However, it does not really matter in this PR, so I will revert it.