@contramundum53 2023-12-27
> The reason why I take `np.size` rather than `len` is that `np.size` imposes a stricter constraint on the input. However, it does not really matter in this PR, so I will revert it.

`np.size` actually doesn't impose constraint on the input. https://numpy.org/doc/stable/reference/generated/numpy.ndarray.size.html

It returns the total number of elements, regardless of the dimension.